+++ src/components/logout/Logout.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+import supabase from "../../config/supabaseClient";
+
+const Logout = async () => {
+  try {
+    const { error } = await supabase.auth.signOut();
+  } catch (error) {
+    alert(error);
+  }
+};
+
+export default Logout;
